Workbook Publish and add/update/delete Permissions Working

+++ scripts/publish_workbook.py
@@ -176,51 +176,49 @@ def main(arguments):
                     "The project_path field is Null in JSON Template.")
             else:
                 # Step: Form a new workbook item and publish.
-                # publish_workbook(server, data)
+                publish_workbook(server, data)
 
                 # Step: Get the Workbook ID from the Workbook Name
                 wb_id = get_workbook_id(server, data)[0]
 
-                # Step: Get the User ID of permission assigned
-                permission_user_id = get_user_id(
-                    server, data['permissions']['permission_user_name'])[0]
+                if "permissions" in data and "permission_template" in data and data['permissions']['permission_template'] is not None:
+                    # Step: Get the User ID of permission assigned
+                    permission_user_id = get_user_id(
+                        server, data['permissions']['permission_user_name'])[0]
 
-                # get permissions of specific workbook
-                user_permissions = query_permission(
-                    data, wb_id, permission_user_id, version, auth_token)
+                    # get permissions of specific workbook
+                    user_permissions = query_permission(
+                        data, wb_id, permission_user_id, version, auth_token)
 
-                for permission_name, permission_mode in data['permissions']['permission_template'].items():
-                    update_permission_flag = True
-                    if user_permissions is None:
-                        add_permission(
-                            data, wb_id, permission_user_id, version, auth_token, permission_name, permission_mode)
-                        print(
-                            f"\tPermission {permission_name} is set to {permission_mode} Successfully in {wb_id}\n")
-                        update_permission_flag = False
-                    else:
-                        for permission in user_permissions:
-                            if permission.get('name') == permission_name:
-                                if permission.get('mode') != permission_mode:
-                                    existing_mode = permission.get('mode')
-                                    delete_permission(
-                                        data, auth_token, wb_id, user_id, permission_name, existing_mode, version)
-                                    update_permission_flag = True
-                                    print(
-                                        f"\tPermission {permission_name} : {existing_mode} is deleted Successfully in {wb_id}\n")
-                                else:
-                                    update_permission_flag = False
+                    for permission_name, permission_mode in data['permissions']['permission_template'].items():
+                        update_permission_flag = True
+                        if user_permissions is None:
+                            add_permission(
+                                data, wb_id, permission_user_id, version, auth_token, permission_name, permission_mode)
+                            print(
+                                f"\tPermission {permission_name} is set to {permission_mode} Successfully in {wb_id}\n")
+                            update_permission_flag = False
+                        else:
+                            for permission in user_permissions:
+                                if permission.get('name') == permission_name:
+                                    if permission.get('mode') != permission_mode:
+                                        existing_mode = permission.get('mode')
+                                        delete_permission(
+                                            data, auth_token, wb_id, user_id, permission_name, existing_mode, version)
+                                        update_permission_flag = True
+                                        print(
+                                            f"\tPermission {permission_name} : {existing_mode} is deleted Successfully in {wb_id}\n")
+                                    else:
+                                        update_permission_flag = False
 
-                    if update_permission_flag:
-                        add_permission(
-                            data, wb_id, user_id, version, auth_token, permission_name, permission_mode)
-                        print(
-                            f"\tPermission {permission_name} is set to {permission_mode} Successfully in {wb_id}\n")
-                    else:
-                        print(
-                            f"\tPermission {permission_name} is already set to {permission_mode} on {data['name']}\n")
-
-                # Step: Update Project permissions
-                # add_permission(data, wb_id, user_id, version, auth_token)
+                        if update_permission_flag:
+                            add_permission(
+                                data, wb_id, user_id, version, auth_token, permission_name, permission_mode)
+                            print(
+                                f"\tPermission {permission_name} is set to {permission_mode} Successfully in {wb_id}\n")
+                        else:
+                            print(
+                                f"\tPermission {permission_name} is already set to {permission_mode} on {data['name']}\n")
 
             # Step: Sign Out to the Tableau Server
             server.auth.sign_out()
